products/tier2/dashboard-analytics/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Database (shared from tier1)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Analytics
pandas==2.1.4
numpy==1.26.2

# Security
python-jose[cryptography]==3.3.0

# Utils
httpx==0.25.2
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

